+++ NEWS.md
@@ -6,6 +6,7 @@
 
 ## Changes to functions
 
+* The `as.tag`-argument in `set_na()` now defaults to `FALSE`.
 * `rec()`, `split_var()` and `dicho()` get an argument `suffix`, to append a suffix to variable (column) names, if applied on a data frame.
 * `find_var()` gets a `as.df`-argument, to return a data frame with matching variables, instead of their column indices only.
 * `find_var()` gets a `as.varlab`-argument, to return a "summary" data frame with column number, variable name and variable label.